Update sketch.js
This code outputs sets of "leaves" which are made up of array of circles, recursive trees, a ground, and a gradient background. All I really have to do left is make it so that the use  can mouse click to erase 1 set of circles at a time (back to front). If I have time, I want to add music and MAYBE make it possible for the user to press a keys on the keyboard to change the sky background (just for fun).