Update sketch.js
This code does MOSTLY what I want it to do. That is, the sky changes colors with keys A,S and D and the circles change colors with mouse movement. However, instead of the layers of circles being "erased" with each mouse click, they just go behind each other. Obviously something needs to be changed on lines 91 to 112, but otherwise everything is looking good! I'll probably have extra time so I'll add some music in the background.